Did all but one clean up suggested by syt0r.
