+++ core/src/commonMain/kotlin/ua/syt0r/kanji/core/PathExtensions.kt
@@ -1,5 +1,6 @@
 package ua.syt0r.kanji.core
 
+import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Path
 import androidx.compose.ui.graphics.PathMeasure
 import kotlin.math.min
@@ -32,11 +33,6 @@ fun Path.approximateEvenly(pointsCount: Int): ApproximatedPath {
     // the end point of the last segment.
     val points = (0 until pointsCount).map {
         val fraction = it.toFloat() / (pointsCount-1) * pathLength
-
-        // points in second "half" were identical which threw off center calculation in
-        // DefaultKanjiStrokeEvaluator
-//        val fraction = it.toFloat() / pointsCount * pathLength +
-//                it.toFloat() / pointsCount * pathLength
         val point = pathMeasure.pointAt(min(fraction, pathLength))
         PathPointF(fraction, point.x, point.y)
     }
@@ -46,6 +42,33 @@ fun Path.approximateEvenly(pointsCount: Int): ApproximatedPath {
         points = points
     )
 }
+
+fun List<PointF>.center(): PointF {
+    return PointF(
+        sumOf { it.x.toDouble() / size }.toFloat(),
+        sumOf { it.y.toDouble() / size }.toFloat()
+    )
+}
+
+fun List<PointF>.minus(value: PointF): List<PointF> {
+    return map { PointF(it.x - value.x, it.y - value.y) }
+}
+
+fun relativeScale(
+    first: List<PointF>,
+    second: List<PointF>
+): Pair<Size, Size> {
+    fun List<PointF>.getScale(): Size = Size(
+        width = run { maxOf { it.x } - minOf { it.x } },
+        height = run { maxOf { it.y } - minOf { it.y } }
+    )
+
+    val firstSize = first.getScale()
+    val secondSize = second.getScale()
+
+    return firstSize to secondSize
+}
+
 
 private const val INTERPOLATION_POINTS = 1 + 195
