+++ core/src/commonMain/kotlin/ua/syt0r/kanji/core/stroke_evaluator/DefaultKanjiStrokeEvaluator.kt
@@ -1,12 +1,15 @@
 package ua.syt0r.kanji.core.stroke_evaluator
 
-import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Path
 import ua.syt0r.kanji.core.PointF
 import ua.syt0r.kanji.core.approximateEvenly
 import ua.syt0r.kanji.core.logger.Logger
 import ua.syt0r.kanji.presentation.common.ui.kanji.KanjiSize
 import kotlin.math.*
+import ua.syt0r.kanji.core.center
+import ua.syt0r.kanji.core.minus
+import ua.syt0r.kanji.core.relativeScale
+
 
 class DefaultKanjiStrokeEvaluator : KanjiStrokeEvaluator {
 
@@ -62,31 +65,31 @@ class DefaultKanjiStrokeEvaluator : KanjiStrokeEvaluator {
         return cumulativeError
     }
 
-    private fun List<PointF>.center(): PointF {
-        return PointF(
-            sumOf { it.x.toDouble() / size }.toFloat(),
-            sumOf { it.y.toDouble() / size }.toFloat()
-        )
-    }
-
-    private fun List<PointF>.minus(value: PointF): List<PointF> {
-        return map { PointF(it.x - value.x, it.y - value.y) }
-    }
-
-    private fun relativeScale(
-        first: List<PointF>,
-        second: List<PointF>
-    ): Pair<Size, Size> {
-        fun List<PointF>.getScale(): Size = Size(
-            width = run { maxOf { it.x } - minOf { it.x } },
-            height = run { maxOf { it.y } - minOf { it.y } }
-        )
-
-        val firstSize = first.getScale()
-        val secondSize = second.getScale()
-
-        return firstSize to secondSize
-    }
+//    private fun List<PointF>.center(): PointF {
+//        return PointF(
+//            sumOf { it.x.toDouble() / size }.toFloat(),
+//            sumOf { it.y.toDouble() / size }.toFloat()
+//        )
+//    }
+//
+//    private fun List<PointF>.minus(value: PointF): List<PointF> {
+//        return map { PointF(it.x - value.x, it.y - value.y) }
+//    }
+//
+//    private fun relativeScale(
+//        first: List<PointF>,
+//        second: List<PointF>
+//    ): Pair<Size, Size> {
+//        fun List<PointF>.getScale(): Size = Size(
+//            width = run { maxOf { it.x } - minOf { it.x } },
+//            height = run { maxOf { it.y } - minOf { it.y } }
+//        )
+//
+//        val firstSize = first.getScale()
+//        val secondSize = second.getScale()
+//
+//        return firstSize to secondSize
+//    }
 
     private fun List<PointF>.scaled(scaleX: Float, scaleY: Float): List<PointF> {
         return map { PointF(it.x * scaleX, it.y * scaleY) }
